// src/app/services/idb/features/features.enum.ts
export enum FeaturesName {
	dashboard = 'Dashboard',
	spm = 'SCRUM PM',
	teams = 'SCRUM Teams',
	planningRoom = 'Planning Room',
	groomingRoom = 'Grooming Room',
	dailyRoom = 'Daily Room',
	reviewRoom = 'Review Room',
	retrospectiveRoom = 'Retrospective Room',
	achievements = 'Achievements',
	history = 'History',
	dataExport = 'Data export',
	autoReport = 'Auto report',
	resources = 'Resources',
	backlogs = 'Backlogs'
}
export enum FeaturesIcon {
	dashboard = 'data_exploration',
	spm = 'manage_history',
	teams = 'groups',
	planningRoom = 'meeting_room',
	groomingRoom = 'grain',
	dailyRoom = 'today',
	reviewRoom = 'previews',
	retrospectiveRoom = 'groups_3',
	achievements = 'emoji_events',
	history = 'history',
	dataExport = 'file_download',
	autoReport = 'smart_toy',
	resources = 'psychology_alt',
	backlogs = 'large_receipe'
}